# By Mikhail Polubisok
# Via Mikhail Polubisok
* commit '272608e8e3591e76873eccf2b18f004e11450f8c':
  [DF-6249] Fix review issues
  [DF-6249] Fix review issues
  [DF-6249] Setup default screenScaleFactor option.
  [DF-6249] Add core option "scaleScreenFactor".

+++ Sources/Internal/Platform/DeviceInfo.h
@@ -44,7 +44,7 @@ public:
 	{
 		int32 width;
 		int32 height;
-		int32 scale;
+		float32 scale;
 		
 		ScreenInfo()
 		{
@@ -53,7 +53,7 @@ public:
 			scale = 1;
 		}
 		
-		ScreenInfo(int32 w, int32 h, int32 _scale)
+		ScreenInfo(int32 w, int32 h, float32 _scale)
 		{
 			width = w;
 			height = h;
